+++ apps/create-kunkun/CHANGELOG.md
@@ -1,5 +1,13 @@
 # create-kunkun
 
+## 0.1.28
+
+### Patch Changes
+
+- Re-enable server-based dev extension refresh. Deep Link forces switch focus to kunkun
+- Updated dependencies
+  - @kksh/api@0.0.21
+
 ## 0.1.27
 
 ### Patch Changes